lib to old, and a new lib file
